Use Gatsby's Img component to display inline images from Drupal.
Adds some code to handle dealing with inline images added to the body of a blog post via a <drupal-media/> token. It require changes to lehub, specifically, this won't work without code in-place on the Drupal side to add additinoal data-* attributes to the HTML provided by the Media module.