Allow flexible version resolution for gatsby-theme-osiolabs-drupal.
This is required to fix a problem where our current setup can result in multiple different versions of gatsby-theme-osiolabs-drupal being installed because without a semver range it'll use HEAD. Which may or may not be the same as what the main project (e.g. heynode.com) resolves too. Using a range like this means we can just use whatever version the parent project requires.